Use flatbuffers for BuildFile
Much better performance than YAML for both emitting and parsing.
flatbuffers build files also tend to be smaller than the equivalent
YAML documents, and unlike the YAML build files they do not require
producers to put fields in a particular order (a requirement that is
incidentally not compliant to the YAML spec).